# feat: change underlying geospatial index to RBush