--- a/.github/workflows/node.yml
+++ b/.github/workflows/node.yml
@@ -1,4 +1,4 @@
-name: Node
+name: Tests
 on: [push, pull_request]
 jobs:
   test: