--- a/package.json
+++ b/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "supercluster",
-  "version": "8.0.1",
-  "description": "A very fast geospatial point clustering library.",
+  "name": "mutable-supercluster",
+  "version": "1.0.0",
+  "description": "A geospatial point clustering library forked from the 'supercluster' library and adapted to be mutable.",
   "main": "dist/supercluster.js",
   "type": "module",
   "exports": "./index.js",
@@ -24,17 +24,17 @@
   ],
   "repository": {
     "type": "git",
-    "url": "git://github.com/mapbox/supercluster.git"
+    "url": "git://github.com/SegmentationFaults0/mutable_supercluster.git"
   },
   "keywords": [
     "clustering",
     "geospatial",
     "markers"
   ],
-  "author": "Vladimir Agafonkin",
+  "author": "Sander Verwimp",
   "license": "ISC",
   "dependencies": {
-    "kdbush": "^4.0.2"
+    "rbush": "^4.0.1"
   },
   "devDependencies": {
     "@rollup/plugin-node-resolve": "^15.2.3",