--- a/test/test.js
+++ b/test/test.js
@@ -5,6 +5,8 @@ import assert from 'node:assert/strict';
 import {readFileSync} from 'fs';
 import Supercluster from '../index.js';
 
+const getId = point => point.pointId;
+
 const compareFn = (a, b) => a.geometry[0][0] - b.geometry[0][0] ||
                             a.geometry[0][1] - b.geometry[0][1] ||
                             a.geometry[0][2] - b.geometry[0][2] ||
@@ -17,27 +19,27 @@ const placesTileMin5 = JSON.parse(readFileSync(new URL('./fixtures/places-z0-0-0
 const sortedTileMin5Features = placesTileMin5.features.sort(compareFn);
 
 test('generates clusters properly', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     const tile = index.getTile(0, 0, 0);
     tile.features.sort(compareFn);
     assert.deepEqual(tile.features, sortedTileFeatures);
 });
 
 test('supports minPoints option', () => {
-    const index = new Supercluster({minPoints: 5}).load(places.features);
+    const index = new Supercluster({minPoints: 5, getId}).load(places.features);
     const tile = index.getTile(0, 0, 0);
     tile.features.sort(compareFn);
     assert.deepEqual(tile.features, sortedTileMin5Features);
 });
 
 test('returns children of a cluster', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     const childCounts = index.getChildren(164).map(p => p.properties.point_count || 1);
     assert.deepEqual(childCounts, [1, 7, 2, 6]);
 });
 
 test('returns leaves of a cluster', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     const leafNames = index.getLeaves(164, 10, 5).map(p => p.properties.name);
     assert.deepEqual(leafNames, [
         'I. de Cozumel',
@@ -54,13 +56,13 @@ test('returns leaves of a cluster', () => {
 });
 
 test('generates unique ids with generateId option', () => {
-    const index = new Supercluster({generateId: true}).load(places.features);
+    const index = new Supercluster({generateId: true, getId}).load(places.features);
     const ids = index.getTile(0, 0, 0).features.filter(f => !f.tags.cluster).map(f => f.id);
     assert.deepEqual(ids, [62,  24, 22,  12, 28, 20, 125, 119, 30, 118, 81, 21, 81, 118]);
 });
 
 test('getLeaves handles null-property features', () => {
-    const index = new Supercluster().load(places.features.concat([{
+    const index = new Supercluster({getId}).load(places.features.concat([{
         type: 'Feature',
         properties: null,
         geometry: {
@@ -73,7 +75,7 @@ test('getLeaves handles null-property features', () => {
 });
 
 test('returns cluster expansion zoom', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     assert.deepEqual(index.getClusterExpansionZoom(164), 1);
     assert.deepEqual(index.getClusterExpansionZoom(196), 1);
     assert.deepEqual(index.getClusterExpansionZoom(581), 2);
@@ -86,6 +88,7 @@ test('returns cluster expansion zoom for maxZoom', () => {
         radius: 60,
         extent: 256,
         maxZoom: 4,
+        getId
     }).load(places.features);
 
     assert.deepEqual(index.getClusterExpansionZoom(2504), 5);
@@ -95,7 +98,8 @@ test('aggregates cluster properties with reduce', () => {
     const index = new Supercluster({
         map: props => ({sum: props.scalerank}),
         reduce: (a, b) => { a.sum += b.sum; },
-        radius: 100
+        radius: 100,
+        getId
     }).load(places.features);
 
     assert.deepEqual(index.getTile(1, 0, 0).features.map(f => f.tags.sum).filter(Boolean),
@@ -105,7 +109,7 @@ test('aggregates cluster properties with reduce', () => {
 });
 
 test('returns clusters when query crosses international dateline', () => {
-    const index = new Supercluster().load([
+    const index = new Supercluster({getId}).load([
         {
             type: 'Feature',
             properties: null,
@@ -146,7 +150,7 @@ test('returns clusters when query crosses international dateline', () => {
 });
 
 test('does not crash on weird bbox values', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     assert.equal(index.getClusters([129.426390, -103.720017, -445.930843, 114.518236], 1).length, 26);
     assert.equal(index.getClusters([112.207836, -84.578666, -463.149397, 120.169159], 1).length, 27);
     assert.equal(index.getClusters([129.886277, -82.332680, -445.470956, 120.390930], 1).length, 26);
@@ -157,7 +161,7 @@ test('does not crash on weird bbox values', () => {
 });
 
 test('does not crash on non-integer zoom values', () => {
-    const index = new Supercluster().load(places.features);
+    const index = new Supercluster({getId}).load(places.features);
     assert.ok(index.getClusters([179, -10, -177, 10], 1.25));
 });
 
@@ -165,7 +169,8 @@ test('makes sure same-location points are clustered', () => {
     const index = new Supercluster({
         maxZoom: 20,
         extent: 8192,
-        radius: 16
+        radius: 16,
+        getId
     }).load([
         {type: 'Feature', geometry: {type: 'Point', coordinates: [-1.426798, 53.943034]}},
         {type: 'Feature', geometry: {type: 'Point', coordinates: [-1.426798, 53.943034]}}
@@ -175,7 +180,7 @@ test('makes sure same-location points are clustered', () => {
 });
 
 test('makes sure unclustered point coords are not rounded', () => {
-    const index = new Supercluster({maxZoom: 19}).load([
+    const index = new Supercluster({maxZoom: 19, getId}).load([
         {type: 'Feature', geometry: {type: 'Point', coordinates: [173.19150559062456, -41.340357424709275]}}
     ]);
 
@@ -184,7 +189,7 @@ test('makes sure unclustered point coords are not rounded', () => {
 
 test('does not throw on zero items', () => {
     assert.doesNotThrow(() => {
-        const index = new Supercluster().load([]);
+        const index = new Supercluster({getId}).load([]);
         assert.deepEqual(index.getClusters([-180, -85, 180, 85], 0), []);
     });
 });