… test/fixtures/places.json