This PR changes the underlying geospatial index of clustering library from a KD-Tree to an R-Tree. This is done as a first step towards making the cluster mutable after it is built. 

Unfortunately, using the new index comes with a cost as well. These are the results form the benchmark that is included in the project:
- Before:
```
prepare 1000000 points: 471.137ms
z6: 230371 clusters in 416ms
z5: 64804 clusters in 150ms
z4: 16648 clusters in 49ms
z3: 4212 clusters in 7ms
z2: 1074 clusters in 2ms
z1: 276 clusters in 0ms
z0: 72 clusters in 0ms
total time: 1.110s
memory used: 66452 KB 
```
- After:
```
prepare 1000000 points: 1.317s
z6: 230065 clusters in 1584ms
z5: 64701 clusters in 312ms
z4: 16633 clusters in 58ms
z3: 4185 clusters in 11ms
z2: 1057 clusters in 3ms
z1: 267 clusters in 1ms
z0: 73 clusters in 1ms
total time: 3.294s
memory used: 205868 KB
```